+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Before every entry put one of these to mark the severity of the change:
 - [Minor] Added `with` statement support to CmmtPositionFunctionBlock.
 ### Fixed
 - [Patch] Corrected exception upon failing telegram assertion.
- 
+- [Patch] Fixed exception caused by reading non-existing DeviceInformationRequest for older CMMT devices.
 ## [v0.4.0] - 24.06.22
 ### Added
 - [Minor] Added CLI options for edcon-position tool.