+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Before every entry put one of these to mark the severity of the change:
 - [Patch] Corrected exception upon failing telegram assertion.
 - [Patch] Fixed exception caused by reading non-existing DeviceInformationRequest for older CMMT devices.
 - [Patch] Fixed bug causing some drives to not start traversing task because of missing delay.
+- [Patch] Fixed exception caused when PNU was accessed via Modbus on legacy devices.
 ## [v0.4.0] - 24.06.22
 ### Added
 - [Minor] Added CLI options for edcon-position tool.