+++ CHANGELOG.md
@@ -9,10 +9,12 @@ Before every entry put one of these to mark the severity of the change:
 ## Unreleased
 ### Added
 - [Minor] Added sphinx docs.
-- [Minor] Added `with` statement support to CmmtPositionFunctionBlock.
+- [Minor] Added `with` statement support to EDrivePositioning.
+### Changed
+- [Minor] Renamed `CMMT` occurences to `EDrive`
 ### Fixed
 - [Patch] Corrected exception upon failing telegram assertion.
-- [Patch] Fixed exception caused by reading non-existing DeviceInformationRequest for older CMMT devices.
+- [Patch] Fixed exception caused by reading non-existing DeviceInformationRequest for older EDrive devices.
 - [Patch] Fixed bug causing some drives to not start traversing task because of missing delay.
 - [Patch] Fixed exception caused when PNU was accessed via Modbus on legacy devices.
 ## [v0.4.0] - 24.06.22