+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Before every entry put one of these to mark the severity of the change:
 - [Patch] Fixed exception caused by reading non-existing DeviceInformationRequest for older EDrive devices.
 - [Patch] Fixed bug causing some drives to not start traversing task because of missing delay.
 - [Patch] Fixed exception caused when PNU was accessed via Modbus on legacy devices.
+- [Patch] Fixed bug that sometimes caused infinite waiting for (homing, position, record) tasks to be finished.
 ## [v0.4.0] - 24.06.22
 ### Added
 - [Minor] Added CLI options for edcon-position tool.